move to shared and split into EtherscanLike and Etherscan
