+++ packages/backend/src/peripherals/aztec/schemas.ts
@@ -1,7 +1,6 @@
+import { stringAs } from '@l2beat/shared'
 import { UnixTime } from '@l2beat/shared-pure'
 import { z } from 'zod'
-
-import { stringAs } from '../../tools/types'
 
 export type Rollup = z.infer<typeof Rollup>
 const mined = stringAs((s) => UnixTime.fromDate(new Date(s)))
