+++ packages/backend/src/peripherals/loopring/schemas.ts
@@ -1,7 +1,6 @@
+import { numberAs } from '@l2beat/shared'
 import { UnixTime } from '@l2beat/shared-pure'
 import { z } from 'zod'
-
-import { numberAs } from '../../tools/types'
 
 const MS_IN_SECOND = 1000
