+++ packages/backend/src/peripherals/zksync/schemas.ts
@@ -1,7 +1,6 @@
+import { stringAs } from '@l2beat/shared'
 import { UnixTime } from '@l2beat/shared-pure'
 import { z } from 'zod'
-
-import { stringAs } from '../../tools/types'
 
 export type ZksyncSuccessResponse = z.infer<typeof ZksyncSuccessResponse>
 const ZksyncSuccessResponse = z.object({
