+++ packages/shared/src/services/etherscan/EtherscanClient.ts
@@ -1,83 +1,56 @@
-import { EthereumAddress, Hash256, RateLimiter } from '@l2beat/shared-pure'
+import {
+  ChainId,
+  EthereumAddress,
+  Hash256,
+  UnixTime,
+} from '@l2beat/shared-pure'
 
+import { Logger } from '../../tools'
+import { stringAsInt } from '../../tools/types'
+import { EtherscanLikeClient } from '../etherscanlike/EtherscanLikeClient'
 import { HttpClient } from '../HttpClient'
+import { BlockNumberProvider } from '../providers/BlockNumberProvider'
 import {
   ContractCreatorAndCreationTxHashResult,
   ContractSourceResult,
-  EtherscanResponse,
 } from './model'
 
-export class EtherscanClient {
-  private readonly rateLimiter = new RateLimiter({
-    callsPerMinute: 120,
-  })
-  private readonly timeoutMs = 10_000
+export class EtherscanError extends Error {}
 
-  constructor(
-    private readonly httpClient: HttpClient,
-    private readonly url: string,
-    private readonly apiKey: string,
-    private readonly retryCount = 3,
-  ) {
-    this.call = this.rateLimiter.wrap(this.call.bind(this))
+export class EtherscanClient
+  extends EtherscanLikeClient
+  implements BlockNumberProvider
+{
+  static API_URL = 'https://api.etherscan.io/api'
+
+  constructor(httpClient: HttpClient, apiKey: string, logger = Logger.SILENT) {
+    super(httpClient, EtherscanClient.API_URL, apiKey, logger)
+  }
+
+  getChainId(): ChainId {
+    return ChainId.ETHEREUM
+  }
+
+  async getBlockNumberAtOrBefore(timestamp: UnixTime): Promise<number> {
+    const result = await this.call('block', 'getblocknobytime', {
+      timestamp: timestamp.toString(),
+      closest: 'before',
+    })
+    return stringAsInt().parse(result)
   }
 
   async getContractSource(address: EthereumAddress) {
     const response = await this.call('contract', 'getsourcecode', {
       address: address.toString(),
     })
-    if (response.status !== '1') {
-      throw new Error(
-        `Error response ${response.message} ${JSON.stringify(response.result)}`,
-      )
-    }
-    return ContractSourceResult.parse(response.result)[0]
+    return ContractSourceResult.parse(response)[0]
   }
 
   async getContractDeploymentTx(address: EthereumAddress): Promise<Hash256> {
     const response = await this.call('contract', 'getcontractcreation', {
       contractaddresses: address.toString(),
     })
-    if (response.status !== '1') {
-      throw new Error(
-        `Error response ${response.message} ${JSON.stringify(response.result)}`,
-      )
-    }
 
-    return ContractCreatorAndCreationTxHashResult.parse(response.result)[0]
-      .txHash
-  }
-
-  async call(module: string, action: string, params: Record<string, string>) {
-    const query = new URLSearchParams({
-      module,
-      action,
-      ...params,
-      apikey: this.apiKey,
-    })
-    const url = `${this.url}?${query.toString()}`
-
-    let res = undefined
-    for (let i = 0; i < this.retryCount; i++) {
-      try {
-        res = await this.httpClient.fetch(url, { timeout: this.timeoutMs })
-        break
-      } catch (err) {
-        continue
-      }
-    }
-
-    if (!res) {
-      throw new Error(`Failed to fetch ${url}`)
-    }
-
-    if (!res.ok) {
-      throw new Error(
-        `Server responded with non-2XX result: ${res.status} ${res.statusText}`,
-      )
-    }
-
-    const json = (await res.json()) as unknown
-    return EtherscanResponse.parse(json)
+    return ContractCreatorAndCreationTxHashResult.parse(response)[0].txHash
   }
 }
